+++ src/chat_view.ts
@@ -12,7 +12,7 @@ import { BoostServiceHelper } from './boostServiceHelper';
 import { boostLogging } from './boostLogging';
 import { BoostNotebook, BoostNotebookCell } from './jupyter_notebook';
 import { ControllerOutputType } from './controllerOutputTypes';
-import { noProjectOpenMessage } from './boostprojectdata_interface';
+import { noProjectOpenMessage, extensionNotFullyActivated, extensionFailedToActivate } from './boostprojectdata_interface';
 
 export const aiName = "Sara";
 
@@ -135,10 +135,20 @@ export class BoostChatViewProvider implements vscode.WebviewViewProvider {
 
         const workspaceFolder = vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders[0] : ""; // Get the first workspace folder
 
-        if (!vscode.workspace.workspaceFolders) {
-            return `<html><body><h1>Boost Sara Chat</h1><p>${noProjectOpenMessage}</p></body></html>`;
-        }
+        let message;
 
+        if (!this._boostExtension.finishedActivation) {
+            message = extensionNotFullyActivated;
+        } else if (!this._boostExtension.successfullyActivated) {
+            message = extensionFailedToActivate;
+        } else if (!vscode.workspace.workspaceFolders) {
+            message = noProjectOpenMessage;
+        }
+        
+        if (message) {
+            return `<html><body><h3>Boost ${aiName} Chat</h3><p>${message}</p></body></html>`;
+        }         
+        
         const projectName = workspaceFolder ? path.basename(workspaceFolder.uri.fsPath) : "your workspace";
 
         const template = _.template(rawHtmlContent);