+++ src/quick_summary_controller.ts
@@ -5,14 +5,14 @@ import {
     markdownMimeType
 } from './base_controller';
 
+import { BoostConfiguration } from './boostConfiguration';
+
 import { buildVSCodeIgnorePattern } from './extension';
 
 import {
     FunctionKernelControllerBase,
 } from './function_base_controller';
 
-
-import { BoostConfiguration } from './boostConfiguration';
 import * as vscode from 'vscode';
 import {
     BoostNotebookCell,
@@ -24,24 +24,32 @@ import { findCellByKernel, generateCellOutputWithHeader } from './extension';
 import { getCurrentOrganization } from "./authorization";
 import { ControllerOutputType } from './controllerOutputTypes';
 
-export const quickComplianceSummaryKernelName = 'quickcompliance';
-const quickComplianceSummaryOutputHeader = `Architectural Quick Summary Compliance`;
+export const quickSummaryKernelNamePrefix = `quick_summary_`;
 
-export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
-	constructor(context: vscode.ExtensionContext, onServiceErrorHandler: any, otherThis : any, collection: vscode.DiagnosticCollection) {
+export class BoostQuickSummaryKernelControllerBase extends KernelControllerBase {
+	constructor(context: vscode.ExtensionContext, onServiceErrorHandler: any, otherThis : any, collection: vscode.DiagnosticCollection,
+            kernelName: string,
+            outputType: ControllerOutputType,
+            coreOutputType: ControllerOutputType,
+            outputHeader: string,
+        ) {
         super(
             collection,
-            quickComplianceSummaryKernelName,
-            'Quick Compliance Report',
-            'Quickly builds a Data Compliance Summary Report from data about project and compliance analysis.',
-            ControllerOutputType.complianceFunction,
-            quickComplianceSummaryOutputHeader,
+            `${quickSummaryKernelNamePrefix}${kernelName}`,
+            `Quick ${outputHeader} Report`,
+            `Quickly builds a ${outputHeader} Summary Report from data about project and ${outputType} analysis.`,
+            outputType,
+            `Architectural Quick Summary ${outputHeader} Report`,
             false,
             false,
             context,
             otherThis,
             onServiceErrorHandler);
-        this._coreOutputType = ControllerOutputType.compliance;
+        this._coreOutputType = coreOutputType;
+
+        // tap the parent controller to get it initialized for any potential config or init errors
+        // NOTE: This assumes the parent has already been initialized
+        this.parentController;
 	}
 
     private _coreOutputType: ControllerOutputType;
@@ -54,7 +62,7 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
         switch (BoostConfiguration.cloudServiceStage)
         {
             case "local":
-                return 'http://127.0.0.1:8000/quick-compliance-summary';
+                return 'http://127.0.0.1:8000/quick-summary';
             case 'dev':
                 return '';
             case "test":
@@ -84,7 +92,7 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
         // are we analyzing a source file or a project?
         let projectWideAnalysis = (notebook.metadata['sourceFile'] as string) === './';
         if (!projectWideAnalysis) {
-            throw new Error("Quick Compliance can only be run at the Project level");
+            throw new Error("Quick Summary can only be run at the Project level");
         }
 
         // now get the current organization
@@ -103,7 +111,7 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
         let successfullyCompleted = true;
         try
         {
-            await this._runQuickCompliance(notebook, authPayload);
+            await this._runQuickSummary(notebook, authPayload);
 
         } catch (rethrow) {
             successfullyCompleted = false;
@@ -130,7 +138,7 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
             if (!(kernelController instanceof FunctionKernelControllerBase)) {
                 continue;
             }
-            if (kernelController.outputType !== this.outputType) {
+            if (kernelController.outputType !== this._coreOutputType) {
                 continue;
             }
             this._parentController = kernelController;
@@ -143,7 +151,7 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
         return this._parentController;
 }
 
-    private async _runQuickCompliance(
+    private async _runQuickSummary(
             notebook: BoostNotebook,
             authPayload: any) {
 
@@ -152,14 +160,14 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
         }
 
         // we don't want to overwrite deep summary, which are far more detailed and useful in general
-        let existingComplianceCell = findCellByKernel(notebook, ControllerOutputType.compliance) as BoostNotebookCell;
-        if (existingComplianceCell && existingComplianceCell.value &&
-            existingComplianceCell.metadata?.summaryType) {
-            if (existingComplianceCell.metadata.summaryType === "summary") {
+        let existingSummaryCell = findCellByKernel(notebook, this.outputType) as BoostNotebookCell;
+        if (existingSummaryCell && existingSummaryCell.value &&
+            existingSummaryCell.metadata?.summaryType) {
+            if (existingSummaryCell.metadata.summaryType === "summary") {
                 boostLogging.info(`Skipping ${this.command} of Project-level Notebook " +
                                   "because it already has a detailed Summary`, false);
                 return;
-            } else if (existingComplianceCell.metadata.summaryType === "quick") {
+            } else if (existingSummaryCell.metadata.summaryType === "quick") {
                 boostLogging.info(`Rebuilding ${this.command} of Project-level Notebook ` +
                                   `from last quick summary`, false);
             }
@@ -223,14 +231,14 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
             }
         }
 
-        let targetCell = findCellByKernel(notebook, this._coreOutputType) as BoostNotebookCell;
+        let targetCell = findCellByKernel(notebook, this.outputType) as BoostNotebookCell;
 
         if (!targetCell) {
             targetCell = new BoostNotebookCell(NotebookCellKind.Markup, "", "markdown");
             targetCell.initializeMetadata(
                 {
                     "id": targetCell.id,
-                    "outputType": this._coreOutputType,
+                    "outputType": this.outputType,
                     "summaryType": "quick"
                 });
             notebook.addCell(targetCell);
@@ -250,10 +258,10 @@ export class BoostQuickComplianceSummaryKernel extends KernelControllerBase {
             //    for the synthesized cell
             const analysisOutput : SerializedNotebookCellOutput = {
                 items: [ { mime: markdownMimeType, data : "" } ],
-                metadata : { outputType: this._coreOutputType,
+                metadata : { outputType: this.outputType,
                     details: tempProcessingCell.outputs[0].metadata?.details } };
 
-            targetCell.updateOutputItem(this._coreOutputType, analysisOutput);
+            targetCell.updateOutputItem(this.outputType, analysisOutput);
         } else {
             targetCell.value = offlineSummary as string;